fixed DOCKER_APP_RELEASE in set-recommended-versions to valid release (#224)
`scripts/set-recommended-versions` has `DOCKER_APP_RELEASE="v0.14"`
which does not seem to exist. This appears to be a similar to the silly
bug described in https://github.com/jippi/docker-pixelfed/issues/105.
Looks like it's been there for a while.

Found it whilst looking for a more simplistic update path for
non-breaking releases and found it was failing due to a non-existent
release defined. Seems like 0.12.5 is appropriate.

---------

Co-authored-by: Christian Winther <jippignu@gmail.com>